“各ファイルへ切り分けた”
