+++ csv_file/setsuzokusi_check.csv
@@ -1,0 +1,1 @@
+けど,だけど,逆に,とはいえ
