+++ document.txt
@@ -1,7 +1,6 @@
-上手くできました。
-9文字
+プログラミングを頑張る。
 -------------------------------------------------
-けど、これからも頑張ります。
-14文字
-●E-2：接続詞が誤っています
+12文字
+●E-3：専門用語が書かれています
+・プログラミング
 -------------------------------------------------
